// src/types/card.ts
export type CardItem = {
    title: string
    description: string
    subdescription?: string
}

export type CardInfos = {
    mainHome: CardItem[]
    mainProduct: CardItem[]
    productCapaTermica: CardItem[]
    productLonaTermica: CardItem[]
    productBolsaTermica: CardItem[]
    productBiPartida: CardItem[]
    productDivisoriaTermica: CardItem[]
    productDivisoriaCortina: CardItem[]
    productGancheira: CardItem[]
    productPrateleira: CardItem[]
}